requirements.txt
absl-py
accelerate
addict
aiofiles
aiohappyeyeballs
aiohttp
aiosignal
albumentations
aliyun-python-sdk-core
aliyun-python-sdk-kms
altair
annotated-types
antlr4-python3-runtime
anyio
async-timeout
attrs
backports.zoneinfo
basicsr
blinker
braceexpand
Brotli
cachetools
certifi 
charset-normalizer
click
contourpy
crcmod
cycler
diffusers
einops
exceptiongroup
fastapi
ffmpy
filelock
fonttools
frozenlist
fsspec
ftfy
future
geographiclib
geopy
gitdb
GitPython
google-auth
google-auth-oauthlib
gradio
grpcio
h11
httpcore
httpx
huggingface-hub
idna 
imageio
imageio-ffmpeg
importlib_metadata
importlib_resources
invisible-watermark
Jinja2
joblib
jsonschema
jsonschema-specifications
keras
kiwisolver
kornia
lazy_loader
linkify-it-py
lmdb
lxml
Markdown
markdown-it-py
MarkupSafe
matplotlib
mdit-py-plugins
mdurl
mkl-fft
mkl-random
mkl-service
mmengine
model-index
mpmath
multidict
narwhals
networkx
numpy
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvtx-cu12
oauthlib
omegaconf
open-clip-torch
opencv-contrib-python
opencv-python
opencv-python-headless
opendatalab
openmim
openxlab
ordered-set
orjson
oss2
packaging
pandas
patsy
pillow 
pkgutil_resolve_name
platformdirs
plotly
prettytable
propcache
protobuf
pyarrow
pyasn1
pyasn1_modules
pycryptodome
pydantic
pydantic_core
pydeck
pyDeprecate
pydub
Pygments
Pympler
pyparsing
PySocks 
python-dateutil
python-multipart
pytorch-lightning
pytz
PyWavelets
PyYAML
qudida
referencing
regex
requests
requests-oauthlib
rich
rpds-py
rsa
safetensors
scikit-image
scikit-learn
scipy
seaborn
semver
six 
smmap
sniffio
starlette
statsmodels
streamlit
streamlit-drawable-canvas
sympy
tabulate
tb-nightly
tenacity
tensorboard
tensorboard-data-server
termcolor
test-tube
threadpoolctl
tifffile
timm
tokenizers
toml
tomli
torch
torchmetrics
torchvision
tornado
tqdm
transformers
triton
typing_extensions
tzdata
tzlocal
uc-micro-py
urllib3
utm
uvicorn
validators
watchdog
wcwidth
webdataset
websockets
Werkzeug
yapf
yarl
zipp
